Updated our files to contain more obvious license.
We still use the BSD license for everything. This has not changed. The endorsement clause has been added to
our copy, but that should not hurt anyone (?). Of course, original copyright for Limechat and associated projects
included within Textual are still documented within Acknowledgments.pdf and Contributors.pdf. We just use a
standard source header now.

If Textual contains work that is not documented or done so improperly, then please let us know as soon as
possible so that it can be fixed.